% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\docType{data}
\name{tbl_sim}
\alias{tbl_sim}
\title{systolic' = 120+(20*scale_this(age)),}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 8 columns.
}
\usage{
tbl_sim
}
\description{
systolic' = 120+(20*scale_this(age)),
}
\keyword{datasets}
